+++ LearnUp/src/Components/Pages/About.jsx
@@ -72,7 +72,6 @@ const Main = styled.main`
   }
 `;
 
-
 const Content = styled.div`
   width: 100%;
   height: 100%;
