--- LearnUp/src/Components/auth.jsx
@@ -1,8 +1,0 @@
-
-
-export const auth = () => {
-  return (
-    <div>auth</div>
-  )
-}
-
